LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/grid-container-expected.txt


FAIL Check that grid items can query grid container assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
